# Revert "Local waker reference optimization"