Reverts rust-lang-nursery/futures-rs#1045

I just did some benchmarking and the old way seems faster

Without reference optimization:
```
test oneshots ... bench:   3,917,664 ns/iter (+/- 364,682)
test oneshots ... bench:   3,853,601 ns/iter (+/- 572,603)
test oneshots ... bench:   3,849,870 ns/iter (+/- 531,131)
```

With reference optimization:
```
test oneshots ... bench:   4,529,390 ns/iter (+/- 949,579)
test oneshots ... bench:   4,454,850 ns/iter (+/- 444,460)
test oneshots ... bench:   4,462,664 ns/iter (+/- 528,435)
```

0.1 (for comparison):
```
test oneshots ... bench:   5,201,813 ns/iter (+/- 824,011)
test oneshots ... bench:   5,433,348 ns/iter (+/- 1,320,998)
test oneshots ... bench:   5,197,584 ns/iter (+/- 1,232,213)
```

Intel(R) Core(TM) i7-4770K CPU @ 3.50GHz, rustc 1.29.0-nightly (6a1c0637c 2018-07-23)

I don't know how representative our benchmark for futures_unordered is, but it consistently shows better numbers for the old approach.