@MajorBreakfast 2018-07-24
@cramertj It seems so odd that the optimized version isn't actually faster. Can we investigate this somehow? Is the benchmark representative of actual use cases?