@Nemo157 2018-07-24
I don't see this performance reduction, I see the reverted benchmarks (bc88ddee) as a couple percent slower than the current `master` (476fc37b):

```
 name                                  476fc37b ns/iter  bc88ddee ns/iter  diff ns/iter  diff %  speedup
 bounded_100_tx                        277,941           286,917                  8,976   3.23%   x 0.97
 bounded_1_tx                          269,190           274,825                  5,635   2.09%   x 0.98
 oneshots                              5,220,307         5,345,811              125,504   2.40%   x 0.98
 task_init                             448,310           449,153                    843   0.19%   x 1.00
 thread_yield_multi_thread             6,660,400         7,202,772              542,372   8.14%   x 0.92
 thread_yield_single_thread_many_wait  2,494,387         2,493,170               -1,217  -0.05%   x 1.00
 thread_yield_single_thread_one_wait   883,939           889,001                  5,062   0.57%   x 0.99
 unbounded_100_tx                      180,915           188,218                  7,303   4.04%   x 0.96
 unbounded_1_tx                        170,901           172,000                  1,099   0.64%   x 0.99
 unbounded_uncontended                 84,668            87,034                   2,366   2.79%   x 0.97
```

```
Linux 4.17.4-1-ARCH #1 SMP PREEMPT Tue Jul 3 15:45:09 UTC 2018 x86_64 GNU/Linux
Intel(R) Core(TM) i5-6200U CPU @ 2.30GHz
rustc 1.29.0-nightly (1ecf6929d 2018-07-16)
```