@MajorBreakfast 2018-07-24
Okay, I conducted another benchmark on our iMac:

Without reference optimization:
```
test oneshots ... bench:   2,176,297 ns/iter (+/- 751,780)
test oneshots ... bench:   2,188,195 ns/iter (+/- 885,129)
test oneshots ... bench:   2,237,180 ns/iter (+/- 513,037)
```

With reference optimization:
```
test oneshots ... bench:   2,079,410 ns/iter (+/- 519,451)
test oneshots ... bench:   2,083,924 ns/iter (+/- 930,020)
test oneshots ... bench:   2,099,870 ns/iter (+/- 920,923)
```

0.1 (for comparison):
```
test oneshots ... bench:   3,994,364 ns/iter (+/- 860,767)
test oneshots ... bench:   3,920,499 ns/iter (+/- 988,049)
test oneshots ... bench:   4,106,117 ns/iter (+/- 1,055,798)
```

```
iMac18,3
Intel Core i5 3,8 GHz 4 cores
```

Consitent performance improvement. No revert recommended.