@MajorBreakfast 2018-07-24
It's slightly faster with the optimization on our other iMac as well. I now feel confident that we should keep it.